+++ src/main/java/com/alura/hotel/model/habitacion/DatosRespuestaHabitacion.java
@@ -1,9 +1,7 @@
 package com.alura.hotel.model.habitacion;
 
-import java.math.BigDecimal;
-
 public record DatosRespuestaHabitacion(Long id, Integer numero, Boolean disponible, String descripcion, Integer numCamas,
-                                       BigDecimal precio, Long idTipoHabitacion) {
+                                       Double precio, Long idTipoHabitacion) {
     public DatosRespuestaHabitacion(Habitacion habitacion) {
         this(habitacion.getId(), habitacion.getNumero(), habitacion.getDisponible(), habitacion.getDescripcion(),
                 habitacion.getNumCamas(), habitacion.getPrecio(), habitacion.getTipoHabitacion().getId());
